requirements.txt
django==1.10.3
django-extensions==1.3.3
django-registration==2.2
south==0.8.4
djrill==0.9
django-localflavor==1.0.0
